* Parallel/Messages.h (SeqDataResponse::m_multiplicity): Reduce the width to uint16_t. KmerData has been modified to guarantee that the coverage of a single k-mer fits in 16 bits.
git-svn-id: https://svn.bcgsc.ca/svn/abyss/trunk@54082 ad02020f-2a26-4133-947e-1829317c1416

+++ Parallel/Messages.h
@@ -170,7 +170,7 @@ class SeqDataResponse : public Message
 		IDType m_group;
 		IDType m_id;
 		ExtensionRecord m_extRecord;
-		uint32_t m_multiplicity;
+		uint16_t m_multiplicity;
 };
 
 class SetBaseMessage : public Message
